Move constructors to companion objects
