+++ common/app/model/dotcomrendering/DotcomRenderingSupportTypes.scala
@@ -1,10 +1,16 @@
 package model.dotcomrendering
 
+import com.gu.contentapi.client.model.v1.{Block => APIBlock}
+import com.gu.contentapi.client.utils.format.ImmersiveDisplay
 import common.commercial.{CommercialProperties, EditionCommercialProperties, PrebidIndexSite}
 import model.dotcomrendering.pageElements.PageElement
+import model.{ArticleDateTimes, ContentPage, GUDateTimeFormatNew}
 import navigation._
-import play.api.libs.functional.syntax.toFunctionalBuilderOps
+import org.joda.time.DateTime
 import play.api.libs.json._
+import play.api.mvc.RequestHeader
+import views.support.{ImgSrc, Item300}
+
 
 // We have introduced our own set of objects for serializing data to the DotComponents API,
 // because we don't want people changing the core frontend models and as a side effect,
@@ -26,6 +32,16 @@ case class Tag(
 
 object Tag {
   implicit val writes = Json.writes[Tag]
+
+  def apply(t: model.Tag): Tag = {
+    Tag(
+      t.id,
+      t.properties.tagType,
+      t.properties.webTitle,
+      t.properties.twitterHandle,
+      t.properties.contributorLargeImagePath.map(src => ImgSrc(src, Item300)),
+    )
+  }
 }
 
 case class Block(
@@ -45,6 +61,60 @@ case class Block(
 object Block {
   implicit val pageElementWrites = PageElement.pageElementWrites
   implicit val writes = Json.writes[Block]
+
+  // TODO simplify date fields when DCR is ready
+  def apply(
+    block: APIBlock,
+    page: ContentPage,
+    shouldAddAffiliateLinks: Boolean,
+    request: RequestHeader,
+    isMainBlock: Boolean,
+    calloutsUrl: Option[String],
+    dateTimes: ArticleDateTimes,
+  ): Block = {
+
+    val content = page.item
+
+    // We are passing through the block data here, not the article
+    // the block dateTime types are used for liveblogs
+    val blockCreatedOn = block.createdDate.map(_.dateTime)
+    val blockCreatedOnDisplay =
+      blockCreatedOn.map(dt => GUDateTimeFormatNew.formatTimeForDisplay(new DateTime(dt), request))
+
+    val blockFirstPublished = block.firstPublishedDate.map(_.dateTime)
+    val blockFirstPublishedDisplay =
+      blockFirstPublished.map(dt => GUDateTimeFormatNew.formatTimeForDisplay(new DateTime(dt), request))
+
+    val blockLastUpdated = block.lastModifiedDate.map(_.dateTime)
+    val blockLastUpdatedDisplay =
+      blockLastUpdated.map(dt => GUDateTimeFormatNew.formatTimeForDisplay(new DateTime(dt), request))
+
+    val displayedDateTimes = ArticleDateTimes.makeDisplayedDateTimesDCR(dateTimes, request)
+    val campaigns = page.getJavascriptConfig.get("campaigns")
+
+    Block(
+      id = block.id,
+      elements = DotcomRenderingUtils.blockElementsToPageElements(
+        block.elements,
+        request,
+        content,
+        shouldAddAffiliateLinks,
+        isMainBlock,
+        content.metadata.format.exists(_.display == ImmersiveDisplay),
+        campaigns,
+        calloutsUrl,
+      ),
+      blockCreatedOn = blockCreatedOn,
+      blockCreatedOnDisplay = blockCreatedOnDisplay,
+      blockLastUpdated = blockLastUpdated,
+      blockLastUpdatedDisplay = blockLastUpdatedDisplay,
+      title = block.title,
+      blockFirstPublished = blockFirstPublished,
+      blockFirstPublishedDisplay = blockFirstPublishedDisplay,
+      primaryDateLine = displayedDateTimes.primaryDateLine,
+      secondaryDateLine = displayedDateTimes.secondaryDateLine,
+    )
+  }
 }
 
 case class Pagination(
